+++ src/cmap.rs
@@ -1,4 +1,4 @@
-use std::ops::Range;
+use core::ops::Range;
 
 use crate::parser::{Stream, FromData, LazyArray, Offset32};
 use crate::{Font, GlyphId, TableName, Result, Error};