+++ src/gdef.rs
@@ -1,5 +1,5 @@
-use std::convert::{TryFrom, TryInto};
-use std::ops::Range;
+use core::convert::{TryFrom, TryInto};
+use core::ops::Range;
 
 use crate::parser::{Stream, FromData};
 use crate::{GlyphId, Font, TableName, Result, Error};