+++ src/parser.rs
@@ -12,7 +12,7 @@ pub trait FromData: Sized {
     /// For example, when you parsing u16, but storing it as u8.
     /// In this case `size_of::<Self>()` == 1, but `FromData::size_of()` == 2.
     fn size_of() -> usize {
-        std::mem::size_of::<Self>()
+        core::mem::size_of::<Self>()
     }
 }
 
@@ -48,7 +48,7 @@ impl FromData for u32 {
 #[derive(Clone, Copy)]
 pub struct LazyArray<'a, T> {
     data: &'a [u8],
-    phantom: std::marker::PhantomData<T>,
+    phantom: core::marker::PhantomData<T>,
 }
 
 impl<'a, T: FromData> LazyArray<'a, T> {
@@ -58,7 +58,7 @@ impl<'a, T: FromData> LazyArray<'a, T> {
 
         LazyArray {
             data,
-            phantom: std::marker::PhantomData,
+            phantom: core::marker::PhantomData,
         }
     }
 
@@ -90,11 +90,11 @@ impl<'a, T: FromData> LazyArray<'a, T> {
 
     #[inline]
     pub fn binary_search_by<F>(&self, mut f: F) -> Option<T>
-        where F: FnMut(&T) -> std::cmp::Ordering
+        where F: FnMut(&T) -> core::cmp::Ordering
     {
         // Based on Rust std implementation.
 
-        use std::cmp::Ordering;
+        use core::cmp::Ordering;
 
         let mut size = self.len();
         if size == 0 {
@@ -120,12 +120,12 @@ impl<'a, T: FromData> LazyArray<'a, T> {
     }
 }
 
-impl<'a, T: FromData + std::fmt::Debug> std::fmt::Debug for LazyArray<'a, T> {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+impl<'a, T: FromData + core::fmt::Debug> core::fmt::Debug for LazyArray<'a, T> {
+    fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
         let array: LazyArrayIter<T> = LazyArrayIter {
             data: self.data,
             offset: 0,
-            phantom: std::marker::PhantomData,
+            phantom: core::marker::PhantomData,
         };
 
         f.debug_list().entries(array).finish()
@@ -141,7 +141,7 @@ impl<'a, T: FromData> IntoIterator for LazyArray<'a, T> {
         LazyArrayIter {
             data: self.data,
             offset: 0,
-            phantom: std::marker::PhantomData,
+            phantom: core::marker::PhantomData,
         }
     }
 }
@@ -150,7 +150,7 @@ impl<'a, T: FromData> IntoIterator for LazyArray<'a, T> {
 pub struct LazyArrayIter<'a, T> {
     data: &'a [u8],
     offset: usize,
-    phantom: std::marker::PhantomData<T>,
+    phantom: core::marker::PhantomData<T>,
 }
 
 impl<'a, T: FromData> Iterator for LazyArrayIter<'a, T> {